requirements.txt
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.1
pytest==7.4.3
great-expectations==0.18.7
pyarrow==14.0.2
